Rearrange build & utility scripts, update to match linting.
