+++ README.md
@@ -38,9 +38,9 @@ Changes that the shim doesn't account for:
 * `createInstance` now defaults to false, instead of true, when calling `load()`.
 
 ### Asset changes
-The expected asset format is now a module based asset that uses ES6 classes, for use with `require()` (publish for CommonJS) or `import()` (publish for ES6). A script has been provided to update v1 assets to the new format - `scripts/assetConversion.js`.
-* CommonJS export: `node node_modules/pixi-animate/assetConversion.js path/to/myFile.js path/to/my2ndFile.js`
-* ES6 export: `node node_modules/pixi-animate/assetConversion.js -e path/to/myFile.js path/to/my2ndFile.js`
+The expected asset format is now a module based asset that uses ES6 classes, for use with `require()` (publish for CommonJS) or `import()` (publish for ES6). A script has been provided to update v1 assets to the new format - `pixi-animate-upgrade` (see `bin/assetConversion.js` for the source).
+* CommonJS export: `node pixi-animate-upgrade path/to/myFile.js path/to/my2ndFile.js`
+* ES6 export: `node pixi-animate-upgrade -e path/to/myFile.js path/to/my2ndFile.js`
 
 Note that this script does not update graphics paths - you will have to go into your `*.shapes.*` files and replace `"c"` with `"cp"`.
