+++ package.json
@@ -55,7 +55,7 @@
     "prebuild": "npm run lint",
     "build": "tsc && rollup -c && npm run build:min && npm run build:dts",
     "build:min": "terser dist/pixi-animate.js -c -m -o dist/pixi-animate.min.js --source-map \"content='dist/pixi-animate.js.map',includeSources=true,filename='dist/pixi-animate.min.js.map',url='pixi-animate.min.js.map'\" --comments \"/pixi-animate.js - /\"",
-    "build:dts": "node build/buildAmbientTypes.js",
+    "build:dts": "node scripts/buildAmbientTypes.js",
     "test": "floss --path tests/index.js",
     "coverage": "nyc npm test",
     "test:debug": "npm test -- --debug",
@@ -75,8 +75,11 @@
     "package.json",
     "README.md",
     "v1-asset-shim.js",
-    "scripts/assetConversion.js"
+    "bin/"
   ],
+  "bin": {
+    "pixi-animate-upgrade": "bin/assetConversion.js"
+  },
   "keywords": [
     "webgl",
     "pixi",
